Windows build fixes

+++ core/src/main/java/io/questdb/std/json/JsonParser.java
@@ -169,7 +169,7 @@ public class JsonParser implements QuietCloseable {
     private static native int queryPointerInt(
             long impl,
             long jsonPtr,
-            int jsonLen,
+            long jsonLen,
             long jsonTailPadding,
             long pointerPtr,
             long pointerLen,
